Make MSVC happy

+++ src/cat_http_content_type.c
@@ -180,7 +180,7 @@ static cat_always_inline cat_bool_t cat_http_parser_multipart_parse_content_type
             size_t read_len = pe - p > (10 - state + mt_start) ? (10 - state + mt_start) : pe - p;
             // printf("readlen: %d\n", read_len);
             memcpy(&parser->multipart.multipart_boundary[state], p, read_len);
-            state += read_len;
+            state += (int) read_len;
             p += read_len;
 
             if (pe - p == 0) {
